refactor(deepseek_nsa): move golden_nsa_fwd to test file per CI_CHECKLIST §23.3
Address fengz72's review feedback (golden should be in test file, not
example file). Aligns with CI_CHECKLIST §23.2/§23.3 file separation spec.

1. Move golden_nsa_fwd definition from example_tilelang_nsa_fwd.py to
   test_example_tilelang_nsa_fwd.py (golden is test logic, belongs in
   test file per §23.3)

2. example __main__ block: replace golden comparison with shape+finite
   verification only (per §23.2 — __main__ calls one case, no golden
   inline). Outputs [SMOKE_OK] + 'Test Passed!'

3. test file: remove 'from example import golden_nsa_fwd', define
   golden_nsa_fwd locally. Update docstring/comments.

Verified: python example_tilelang_nsa_fwd.py -> [SMOKE_OK] + Test Passed!
         python test_example_tilelang_nsa_fwd.py --level all -> Test Passed (exit 0)
ruff format/check: clean.
